rename and refactor news pydantic models
